move photo validation outside from the PhotoRow component
