+++ src/app/Scenes/Consignments/Screens/SubmitArtworkOverview/UploadPhotos/utils/calculatePhotoSize.ts
@@ -1,5 +1,5 @@
+import { Photo } from "app/Scenes/Consignments/Screens/SubmitArtworkOverview/UploadPhotos/validation"
 import { transformBytesToSize } from "app/utils/transformBytesToSize"
-import { Photo } from "palette/elements/PhotoRow/validation"
 
 const totalSizeLimitInBytes = 30000000
