+++ src/app/Scenes/Consignments/Screens/SubmitArtworkOverview/UploadPhotos/validation.ts
